Remove physics library to use Bullet directly in PhysicsManager and get an implementation going.
