+++ src/CMakeLists.txt
@@ -5,4 +5,3 @@ add_subdirectory(Video)
 add_subdirectory(Engine)
 add_subdirectory(Editor)
 add_subdirectory(Game)
-add_subdirectory(Physics)
