--- src/Physics/README.md
@@ -1,6 +1,0 @@
-# Physics
-Performs physics simulations using Bullet.
-
-## Dependencies
-### External libraries
-- Bullet
